requirements.txt
Pillow
FreeSimpleGUI